powerview refactor
